+++ CHANGELOG.md
@@ -31,6 +31,7 @@
 - Fixed auto backup, auto extension update, and app update checker stop working
   if it crash/failed
 - Fixed crashes when trying to reload extension repo due to connection issue
+- Fixed incompatibility with J2K backup file
 
 ## Translation
 - Update Japanese translation (akir45)